split up configs, disable flash keys
