+++ lua/custom/plugins/flash.lua
@@ -2,7 +2,14 @@ return {
   'folke/flash.nvim',
   event = 'VeryLazy',
   config = function()
-    require('flash').setup()
+    require('flash').setup {
+      modes = {
+        char = {
+          enabled = false,
+          -- keys = { "f", "F", "t", "T", ";", "," },
+        },
+      },
+    }
     vim.keymap.set('n', 'f', '<cmd>lua require("flash").jump()<CR>', { noremap = true, desc = 'flash search' })
     vim.keymap.set(
       'n',
